LPS-53238 Move to separate upgrade process because it was added to UpgradeProcess_1_0_0 after its original introduction, so its possible that some users ran the UpgradeProcess_1_0_0 without UpgradePortletPreferences, in which case UpgradePortletPreferences will never run
