+++ portlets/calendar-portlet/docroot/WEB-INF/src/portal.properties
@@ -1,8 +1,9 @@
-release.info.build.number=100
+release.info.build.number=101
 release.info.previous.build.number=0
 
 upgrade.processes=\
-    com.liferay.calendar.hook.upgrade.UpgradeProcess_1_0_0
+    com.liferay.calendar.hook.upgrade.UpgradeProcess_1_0_0,\
+    com.liferay.calendar.hook.upgrade.UpgradeProcess_1_0_1
 
 portlet.add.default.resource.check.whitelist=1_WAR_calendarportlet
